% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCC.R
\name{TimeCC}
\alias{TimeCC}
\title{Time Cross Correlation Function}
\usage{
TimeCC(x, y)
}
\arguments{
\item{x}{The first time series.}

\item{y}{The second time series.}
}
\value{
A numeric vector representing the combined cross-correlation values for positive and negative lags.
}
\description{
The TimeCC function calculates the cross-correlation between two time series, providing an asymptotically unbiased estimator. The function assumes that the input time series (x and y) are of the same length.
}
\details{
The cross-correlation is not symmetric, so be careful when dealing with negative lags.
}
\examples{
# Example usage:
x <- c(1, 2, 3, 4, 5)
y <- c(5, 4, 3, 2, 1)
result <- TimeCC(x, y)

}
\references{
Insert any relevant references or citations here.
}
\seealso{
\code{\link{crossprod}}, \code{\link{rev}}
}
