% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSPCC.R
\name{varSPCC}
\alias{varSPCC}
\title{Variance of the Sample Pearson Correlation Coefficient}
\usage{
varSPCC(x, y)
}
\arguments{
\item{x}{The first time series.}

\item{y}{The second time series.}
}
\value{
A numeric value representing the variance of the SPCC.
}
\description{
The `varSPCC` function calculates the variance of the Sample Pearson Correlation Coefficient
(SPCC) between two time series, x and y. The function checks if the input time series have the same length and then performs calculations to estimate the variance of the SPCC.
}
\details{
The function estimates the time series correlation (rx and ry) using the `TimeAC` function and the time series cross-correlation (rxy and ryx) using the `TimeCC` function. The variance of the SPCC is then calculated based on these values.
}
\examples{
# Example usage:
x <- c(1, 2, 3, 4, 5)
y <- c(5, 4, 3, 2, 1)
result <- varSPCC(x, y)

}
\references{
Insert any relevant references or citations here.
}
\seealso{
\code{\link{TimeAC}}, \code{\link{TimeCC}}
}
